Clean up /edx/var/edxapp/staticfiles before paver runs
When running on a dirty image (such as a sandbox base image or an AMI
build) this clears static files so that collectstatic will copy
everything.

Django's collectstatic copies files and doesn't preserve timestamps,
then only copies "newer modified files".  This leads to a condition
where old static files are in /edx/var/edxapp/staticfiles with a
timestamp of Yesterday, and we add a new dep which comes down from npm
with a timestamp of last week.  Collectstatic will not copy those.